backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.9
redis==5.0.3
httpx==0.27.0
openai==1.14.3
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic[email]==2.6.4
pytest==8.1.1
pytest-asyncio==0.23.5
